keg: move os specific methods to `os/mac` folder
Also tag certain methods as private.
